refactor: MatchResult가 profitRate알도록 구현 변경
