+++ src/main/java/blackjack/model/betting/Profit.java
@@ -1,6 +1,5 @@
 package blackjack.model.betting;
 
-import blackjack.model.MatchResult;
 import java.util.Objects;
 
 public class Profit {
@@ -13,16 +12,7 @@ public class Profit {
 
     public static Profit of(BetAmount betAmount, MatchResult matchResult) {
         int stake = betAmount.getStake();
-        if (matchResult == MatchResult.BLACKJACK) {
-            return new Profit((int) (1.5 * stake));
-        }
-        if (matchResult == MatchResult.WIN) {
-            return new Profit(stake);
-        }
-        if (matchResult == MatchResult.DRAW) {
-            return new Profit(0);
-        }
-        return new Profit(stake * -1);
+        return new Profit((int)(stake * matchResult.getProfitRate()));
     }
 
     @Override
