+++ src/main/java/blackjack/view/OutputView.java
@@ -2,7 +2,7 @@ package blackjack.view;
 
 import static blackjack.model.constants.RuleConstants.DEALER_HIT_THRESHOLD;
 
-import blackjack.model.MatchResult;
+import blackjack.model.betting.MatchResult;
 import blackjack.model.betting.Profit;
 import blackjack.model.card.Card;
 import blackjack.model.participant.Player;
@@ -49,42 +49,6 @@ public class OutputView {
             return;
         }
         System.out.println();
-    }
-
-    public void printMatchResult(Map<Player, MatchResult> playerMatchResults) {
-        System.out.println("## 최종 승패");
-        printDealerMatchResult(playerMatchResults);
-        printPlayersMatchResult(playerMatchResults);
-    }
-
-    private void printDealerMatchResult(Map<Player, MatchResult> playerMatchResults) {
-        EnumMap<MatchResult, Integer> dealerMatchResult = calculateDealerMatchResult(playerMatchResults);
-        System.out.printf("딜러: %s%n", formatDealerMatchResult(dealerMatchResult));
-    }
-
-    private String formatDealerMatchResult(EnumMap<MatchResult, Integer> dealerMatchResult) {
-        return dealerMatchResult.entrySet()
-                .stream()
-                .filter(entry -> entry.getValue() != 0)
-                .map(entry -> entry.getValue() + entry.getKey().getLabel())
-                .collect(Collectors.joining(" "));
-    }
-
-    private EnumMap<MatchResult, Integer> calculateDealerMatchResult(
-            Map<Player, MatchResult> playerMatchResults) {
-        EnumMap<MatchResult, Integer> dealerMatchResult = new EnumMap<>(MatchResult.class);
-        for (MatchResult matchResult : playerMatchResults.values()) {
-            dealerMatchResult.merge(MatchResult.reverse(matchResult), 1, Integer::sum);
-        }
-        return dealerMatchResult;
-    }
-
-    private void printPlayersMatchResult(Map<Player, MatchResult> playerMatchResults) {
-        for (Entry<Player, MatchResult> playerMatchResultEntry : playerMatchResults.entrySet()) {
-            System.out.printf("%s: %s%n",
-                    playerMatchResultEntry.getKey().getName(),
-                    playerMatchResultEntry.getValue().getLabel());
-        }
     }
 
     public void printPlayerHandAndTotal(List<Player> players) {
