+++ src/test/java/blackjack/model/betting/ProfitTest.java
@@ -2,7 +2,6 @@ package blackjack.model.betting;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.model.MatchResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
